+++ homework1/src/main/java/com/github/patbattb/hw1/commands/InfoMonthlyCommand.java
@@ -1,13 +1,13 @@
 package com.github.patbattb.hw1.commands;
 
-import com.github.patbattb.hw1.service.ReportHandler;
+import com.github.patbattb.hw1.domain.ReportStorage;
 
 /**
  * Prints monthly information.
  */
 public final class InfoMonthlyCommand implements MenuCommand {
     @Override
-    public void runCommand(ReportHandler reportHandler) {
+    public void runCommand(ReportStorage reportStorage) {
 
     }
 }