+++ homework1/src/main/java/com/github/patbattb/hw1/commands/MenuCommand.java
@@ -1,10 +1,10 @@
 package com.github.patbattb.hw1.commands;
 
-import com.github.patbattb.hw1.service.ReportHandler;
+import com.github.patbattb.hw1.domain.ReportStorage;
 
 /**
  * Command interface.
  */
 public interface MenuCommand {
-    void runCommand(ReportHandler reportHandler);
+    void runCommand(ReportStorage reportStorage);
 }