+++ homework1/src/main/java/com/github/patbattb/hw1/domain/CommandMap.java
@@ -1,7 +1,6 @@
 package com.github.patbattb.hw1.domain;
 
 import com.github.patbattb.hw1.commands.*;
-import com.github.patbattb.hw1.service.ReportHandler;
 import com.google.common.collect.ImmutableMap;
 
 /**
@@ -16,7 +15,7 @@ public final class CommandMap {
         mapOfCommands = ImmutableMap.<String, MenuCommand>builder()
                 .put(MenuOptions.READ_MONTHLY_REPORTS.getNumber(), new ReadMonthlyReportsCommand())
                 .put(MenuOptions.READ_YEARLY_REPORTS.getNumber(), new ReadYearlyReportsCommand())
-                .put(MenuOptions.COMPARE_REPORTS.getNumber(), new CompareReportsCommand())
+                .put(MenuOptions.COMPARE_REPORTS.getNumber(), new VerifyReportsCommand())
                 .put(MenuOptions.INFO_MONTHLY.getNumber(), new InfoMonthlyCommand())
                 .put(MenuOptions.INFO_YEARLY.getNumber(), new InfoYearlyCopmmand())
                 .put(MenuOptions.EXIT.getNumber(), new ExitCommand())
@@ -25,7 +24,7 @@ public final class CommandMap {
         defaultCommand = new UnknownCommand();
     }
 
-    public void execute(String number, ReportHandler reportHandler) {
-        mapOfCommands.getOrDefault(number, defaultCommand).runCommand(reportHandler);
+    public void execute(String number, ReportStorage reportStorage) {
+        mapOfCommands.getOrDefault(number, defaultCommand).runCommand(reportStorage);
     }
 }