+++ homework1/src/main/java/com/github/patbattb/hw1/domain/ReportStorage.java
@@ -1,0 +1,22 @@
+package com.github.patbattb.hw1.domain;
+
+/**
+ * Service for handling reports.
+ */
+public final class ReportStorage {
+    private final ReportContainer reportContainer;
+    private final SumContainer sumContainer;
+
+    public ReportStorage() {
+        reportContainer = new ReportContainer();
+        sumContainer = new SumContainer();
+    }
+
+    public ReportContainer getReportContainer() {
+        return reportContainer;
+    }
+
+    public SumContainer getSumContainer() {
+        return sumContainer;
+    }
+}